refactor: IntroActivity 교체
